+++ VpnHood.Tunneling/Messaging/RequestCode.cs
@@ -8,6 +8,7 @@
         TcpDatagramChannel = 2, // data: [4B_jsonLength][json_TcpDatagramChannelRequest]
         TcpProxyChannel = 3, // data: [4B_jsonLength][json_TcpProxyChannelRequest]
         // SessionStatus = 4, // data: [4B_jsonLength][json_BaseRequest]
-        UdpChannel = 5 // data: [4B_jsonLength][json_BaseRequest]
+        UdpChannel = 5, // data: [4B_jsonLength][json_BaseRequest]
+        Bye = 50, // data: [4B_jsonLength][json_BaseRequest]
     }
 }